Refactored TransientException
git-svn-id: https://opencadc.googlecode.com/svn/trunk@1754 728ff76a-78ac-11de-a72b-d90af8dea425

+++ projects/cadcUtil/src/ca/nrc/cadc/net/HttpTransfer.java
@@ -386,7 +386,7 @@ public abstract class HttpTransfer implements Runnable
             return;
         
         boolean trans = false;
-        long dt = 0;
+        int dt = 0;
 
         // try to get the retry delay from the response
         if (code == HttpURLConnection.HTTP_UNAVAILABLE)
@@ -398,7 +398,7 @@ public abstract class HttpTransfer implements Runnable
             {
                 try
                 {
-                    dt = Long.parseLong(retryAfter);
+                    dt = Integer.parseInt(retryAfter);
                     trans = true; // retryReason==SERVER satisfied
                     if (dt > MAX_RETRY_DELAY)
                         dt = MAX_RETRY_DELAY;
